Adjust binstubs for Docker or natively on OS X
Problem
=======

On OS X, we need to initialize the Docker environment
with `docker-machine`.

Within a Docker container,
the `docker-machine` command does not exist.
It is not necessary to do setup through the tool.

Solution
========

Check to see if `docker-machine` exists.

If it exists, run `docker-machine env default`
to set the required environment variables.

If it doesn't exist, then we're good to go.