Consolidate config files into single directory
This directory gets added to each docker image on build.

It is also mounted into the docker containers by the binstubs,
to allow for editing the files from within Docker containers.

To properly mount the `config` directory,
we need to create a base-level `docker-compose.yml`
to define the new Docker build context.

We've added a `docker-compose` binstub
so it doesn't need to be installed on host computers.

`docker-compose` has the benefit of simplifying the build process.
We can now use `docker-compose build`
to rebuild all of our development environment images.